refactor(api)!: gather the public surface behind rekordbox_edit.api
The errors a command raises on purpose now all live in
rekordbox_edit.errors, which docs/api.md documents as the error surface,
rather than three of them hiding in modules that are now private.
ConvertProgress appears in convert()'s signature, so the package
re-exports it, and remove() joins the package's own re-exports.

The API reference gains a Progress section for ConvertProgress, moves
Errors below what it serves, and renders each mkdocstrings block one
level under its section heading instead of beside it.

BREAKING CHANGE: ConvertAborted, ImportInputError, and
DirectoryConfirmationRequired import from rekordbox_edit.errors.
