+++ rekordbox_edit/__init__.py
@@ -4,6 +4,6 @@ __version__ = "0.1.0"
 __author__ = "James Viall"
 __email__ = "jamesviall@pm.me"
 
-from rekordbox_edit.api import convert, edit, import_tracks, search
+from rekordbox_edit.api import convert, edit, import_tracks, remove, search
 
-__all__ = ["search", "edit", "convert", "import_tracks"]
+__all__ = ["search", "edit", "convert", "import_tracks", "remove"]
