+++ rekordbox_edit/api/__init__.py
@@ -4,10 +4,17 @@ Usage:
     from rekordbox_edit.api import search, edit, convert, import_tracks, remove
 """
 
-from rekordbox_edit.api._convert import convert
+from rekordbox_edit.api._convert import ConvertProgress, convert
 from rekordbox_edit.api._edit import edit
 from rekordbox_edit.api._import import import_tracks
 from rekordbox_edit.api._remove import remove
 from rekordbox_edit.api._search import search
 
-__all__ = ["search", "edit", "convert", "import_tracks", "remove"]
+__all__ = [
+    "search",
+    "edit",
+    "convert",
+    "import_tracks",
+    "remove",
+    "ConvertProgress",
+]
